Refactored ThermoPhase::reportCSV
The reportCSV method is responsible for formatting and writing the data provided
by the getCsvReportData method. Derived classes need only override the latter
method to customize their reports.

Also, the data which was not in a CSV format has been removed from the output so
the name of the method is no longer misleading.