[Test] Use more stable output for ChemEquil_ionizedGas
The test output previously included the nondimensional chemical potential
for each species. This contains a term equal to log(mole fraction). In
some of the cases in this test, there are species with very small mass
fractions, below the tolerances of the equilibrium solver, leading to
large apparent changes in the chemical potential for species which are
present in insignificant amounts.

The comparison of the equilibrium mole fractions in this test is retained
by using the "table.csv" file.