[ck2cti] Consider explicit third bodies in duplicate reaction check
Reactions with the same reactants and products but different named
third bodies are not considered duplicates and do not need to be marked
as such.

Normal third body reactions (with 'M' as the third body) are considered
duplicates of reactions with specific third bodies if the efficiency
for the specific third body is not zero.