Reduce confusion related to Fortran 90 sample Makefile
For some compilers (gfortran), there is no space in the argument for specifying
a module directory, e.g. '-Jdir', while for others (ifort), there is,
e.g. '-module dir'. This substitutes the correct prefix directly into the
compiler command line to avoid the confusion associated with having a variable
defined with a meaningful trailing space.