[Kinetics] Fix calculation of coverage-dependent reaction rates
The formula used was not consistent with the standard formulation for
coverage-dependent reaction rates, which uses 10 as the base for the
temperature-independent term with the coverage in the exponential (see
Eq. 11.113 of Kee et al.).

The impact of this change should be minor, as most coverage-dependent reaction
rates appear not to use this term of the parameterization.